refactor ThorExtend
