+++ bin/dockerize-stack_test.rb
@@ -1,7 +1,6 @@
 require 'dockerize_stack'
-require_relative '../lib/rails/dockerize_rails'
 
-DockerizeRails::WORKDIR = "./examples/rails"
+DockerizeRails::WORKDIR = './examples/rails'
 DockerizeStack::Command.start
 
 # ruby bin/dockerize-stack_test.rb rails
