+++ examples/rails/docker/development/entrypoint.sh
@@ -1,0 +1,6 @@
+bundle check || bundle install
+
+rake db:create db:migrate
+rm -f tmp/pids/server.pid
+# rm -rf public/packs/* || true && bin/webpack-dev-server &
+bundle exec rails server -p 3000 -b 0.0.0.0
