+++ examples/rails/docker/kubernetes/scripts/get-secrets.sh
@@ -1,0 +1,1 @@
+kubectl get secrets rails-secrets -o yaml
