+++ examples/rails/docker/kubernetes/scripts/set-secrets.sh
@@ -1,0 +1,5 @@
+kubectl create secret generic rails-secrets \
+  --from-literal=DATABASE_USERNAME= \
+  --from-literal=DATABASE_PASSWORD= \
+  --from-literal=SECRET_KEY_BASE= \
+  --from-literal=RAILS_MASTER_KEY=
