+++ examples/rails/docker/kubernetes/scripts/set-services.sh
@@ -1,0 +1,2 @@
+kubectl apply -f docker/kubernetes/services/cloudsql-proxy.yaml
+kubectl apply -f docker/kubernetes/services/rails-nginx.yaml
