+++ examples/rails/docker/kubernetes/scripts/set-workloads.sh
@@ -1,0 +1,2 @@
+kubectl apply -f docker/kubernetes/workloads/rails-nginx.yaml
+kubectl apply -f docker/kubernetes/workloads/cloudsql-proxy.yaml
