+++ examples/rails/docker/kubernetes/services/rails-nginx.yaml
@@ -1,0 +1,14 @@
+kind: Service
+apiVersion: v1
+metadata:
+  name: rails-nginx
+spec:
+  type: LoadBalancer
+  ports:
+  - port: 80
+    protocol: TCP
+    targetPort: 80
+  selector:
+    app: rails-nginx
+status:
+  loadBalancer: {}
