+++ examples/rails/docker/production/nginx/entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+set -ex
+nginx -g 'daemon off;'
