+++ examples/rails/docker/production/nginx/etc/nginx/conf.d/default.conf
@@ -1,0 +1,34 @@
+server {
+
+   listen 80;
+    server_name localhost;
+    root   /app/public;
+
+    location / {
+      try_files $uri @proxy;
+    }
+
+    location = /favicon.ico {
+      log_not_found off;
+      access_log off;
+    }
+
+    # Block access to "hidden" files and directories whose names begin with a
+    # period. This includes directories used by version control systems such
+    # as Subversion or Git to store control files.
+    location ~ (^|/)\. {
+      return 403;
+    }
+
+    location ~* \.(js|css|png|jpg|jpeg|gif|ico)$ {
+      expires max;
+      log_not_found off;
+    }
+
+    location @proxy {
+      proxy_pass          http://localhost:3000;
+      proxy_set_header    Host             $host;
+      proxy_set_header    X-Real-IP        $remote_addr;
+      proxy_set_header    X-Forwarded-For  $proxy_add_x_forwarded_for;
+    }
+}
