+++ lib/dockerize_stack/dockerize_react.rb
